+++ config/sparktech-auth.php
@@ -6,6 +6,8 @@ return [
 
     'prefix' => env('SPARKTECH_AUTH_PREFIX', 'auth'),
 
+    'api_prefix' => env('SPARKTECH_AUTH_API_PREFIX', 'api'),
+
     'middleware' => ['api'],
 
     'user_model' => App\Models\User::class,